requirements.txt
Django==3.1.1
djangorestframework == 3.11.0
markdown==3.2.2
django-filter==2.3.0
mysqlclient==2.0.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.9
bcrypt==3.2.0
django-cors-headers==3.5.0 
random==2.2.0